Correct arguments in Tornado get_arguments() usage
This method returns an empty list by default, and the second parameter,
'strip,' is a flag to enable stripping of the result strings. In recent
versions of Tornado (4.2.1 here) the 'strip' parameter is checked with
an assert to ensure that its type is Bool, and this Raises in
circus-web.

Previously, the empty list [] would have evaluated Falsey, so replacing
the second parameter with strip=False maintains the same behavior, while
fixing the call in newer versions of Tornado.

Also see
https://github.com/circus-tent/circus-web/issues/59